requirements.txt
fastapi==0.115.12
pydantic==2.10.6
python-dotenv==1.0.1
SQLAlchemy==2.0.39
sqlmodel==0.0.24
fastapi[standard]
aiomysql
pyjwt
passlib[bcrypt]